# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Auto Ecole documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/salles": {
            "get": {
                "tags": [
                    "Salle"
                ],
                "summary": "Get a list of salles",
                "operationId": "aeeed0660e308522aeb9b85c58e43291",
                "responses": {
                    "200": {
                        "description": "List of salles"
                    }
                }
            },
            "post": {
                "tags": [
                    "Salle"
                ],
                "summary": "POST to create salles",
                "operationId": "a6e403cab143ca121bc3a5d608639d49",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "nombreplace": {
                                        "type": "integer"
                                    },
                                    "agence_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create salles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/salles/{id}": {
            "put": {
                "tags": [
                    "Salle"
                ],
                "summary": "Update an salles",
                "operationId": "1e02a561c162278cf40c4d2c0b46232c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the salles to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "nombreplace": {
                                        "type": "integer"
                                    },
                                    "agence_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Salle updated successfully"
                    },
                    "404": {
                        "description": "Salle not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Salle"
                ],
                "summary": "Delete an salles",
                "operationId": "ea52505bdaa729bc8b4487ca37aa9bbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the salles to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "salles deleted successfully"
                    },
                    "404": {
                        "description": "salles not found"
                    }
                }
            }
        },
        "/api/voitures": {
            "get": {
                "tags": [
                    "Voiture"
                ],
                "summary": "Get a list of voitures",
                "operationId": "d30e08bf59ac622fe14c7362320dc603",
                "responses": {
                    "200": {
                        "description": "List of voitures"
                    }
                }
            },
            "post": {
                "tags": [
                    "Voiture"
                ],
                "summary": "POST to create voitures",
                "operationId": "e2a7c92656a073701b2eb4c3bceb4c1e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "nombreplace": {
                                        "type": "integer"
                                    },
                                    "matricule": {
                                        "type": "string"
                                    },
                                    "marque": {
                                        "type": "string"
                                    },
                                    "couleur": {
                                        "type": "string"
                                    },
                                    "agence_id": {
                                        "type": "integer"
                                    },
                                    "personnel_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create voitures",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/voitures/{id}": {
            "put": {
                "tags": [
                    "Voiture"
                ],
                "summary": "Update an voitures",
                "operationId": "1bc82e701dee77e4a8493f97a9271c2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the voitures to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "nombreplace": {
                                        "type": "integer"
                                    },
                                    "matricule": {
                                        "type": "string"
                                    },
                                    "marque": {
                                        "type": "string"
                                    },
                                    "couleur": {
                                        "type": "string"
                                    },
                                    "agence_id": {
                                        "type": "integer"
                                    },
                                    "personnel_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Voiture updated successfully"
                    },
                    "404": {
                        "description": "Voiture not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Voiture"
                ],
                "summary": "Delete an voitures",
                "operationId": "069aac1b335836053a0aeff432fd44cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the voitures to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "voitures deleted successfully"
                    },
                    "404": {
                        "description": "voitures not found"
                    }
                }
            }
        }
    }
}